Step 3 of examples cleanup. Fixing imports and more renaming.
