--- examples/image_classification/README_old.md
@@ -1,9 +1,0 @@
-[home](../../README.md) > [examples](../README.md) > classify
-
-# Examples
-
-This directory contains classification examples.
-
-See:
-* [Image Classification](img/README.md)
-* [Semi-Supervised Image Classification](semi_supervised/img/README.md)
