+++ examples/image_classification/horses_logistic.py
