--- examples/test_generation/README.md
@@ -1,8 +1,0 @@
-[home](../../README.md) > [examples](../README.md) > classify
-
-# Examples
-
-This directory contains examples using RNNs.
-
-See:
-* `shakespeare.py` - predict characters from Shakespeare's plays.
