+++ examples/text_generation/README.md
@@ -1,0 +1,8 @@
+[home](../../README.md) > [examples](../README.md) > text_generation
+
+# Examples
+
+This directory contains text generation examples.
+
+See:
+* `shakespeare_rnn.py` - predict characters from Shakespeare's plays using an RNN.
