+++ examples/text_generation/shakespeare_rnn.py
